+++ internal/controller/cluster_controller.go
@@ -364,10 +364,9 @@ func (r *ClusterMonitoringReconciler) reconcileAlloyServices(ctx context.Context
 		}
 	}
 
-	// Logging-specific: Alloy logs configuration - daemonset that collects data from each node
-	if r.Config.Logging.IsLoggingEnabled(cluster) {
+	// alloy-logs specific: daemonset alloy config, that collects data from each node (not only logs) - TODO rename alloy-logs to alloy-node
+	if r.Config.Logging.IsLoggingEnabled(cluster) || r.Config.Monitoring.IsNetworkMonitoringEnabled(cluster) {
 		// Create or update Alloy logs configuration
-		// TODO make sure we can enable network monitoring separately from logging
 		err = r.AlloyLogsService.ReconcileCreate(ctx, cluster, observabilityBundleVersion)
 		if err != nil {
 			logger.Error(err, "failed to create or update alloy logs config")
@@ -375,7 +374,6 @@ func (r *ClusterMonitoringReconciler) reconcileAlloyServices(ctx context.Context
 		}
 	} else {
 		// Clean up any existing alloy logs configuration
-		// TODO make sure we can enable network monitoring separately from logging
 		err = r.AlloyLogsService.ReconcileDelete(ctx, cluster)
 		if err != nil {
 			logger.Error(err, "failed to delete alloy logs config")
@@ -383,7 +381,7 @@ func (r *ClusterMonitoringReconciler) reconcileAlloyServices(ctx context.Context
 		}
 	}
 
-	// Events-specific: Alloy events configuration - deployment that handles both kube event logs and traces
+	// alloy-event specific: Alloy events configuration - deployment that handles both kube event logs and traces - TODO rename alloy-events to alloy-cluster
 	if r.Config.Logging.IsLoggingEnabled(cluster) || r.Config.Tracing.IsTracingEnabled(cluster) {
 		// Create or update Alloy events configuration
 		err = r.AlloyEventsService.ReconcileCreate(ctx, cluster, observabilityBundleVersion)
@@ -433,10 +431,9 @@ func (r *ClusterMonitoringReconciler) reconcileDelete(ctx context.Context, clust
 			}
 		}
 
-		// Logging-specific: Alloy logs configuration
-		if r.Config.Logging.IsLoggingEnabled(cluster) {
+		// alloy-logs specific: daemonset alloy config, that collects data from each node (not only logs) - TODO rename alloy-logs to alloy-daemonset
+		if r.Config.Logging.IsLoggingEnabled(cluster) || r.Config.Monitoring.IsNetworkMonitoringEnabled(cluster) {
 			// Clean up any existing alloy logs configuration
-			// TODO make sure we can enable network monitoring separately from logging
 			err = r.AlloyLogsService.ReconcileDelete(ctx, cluster)
 			if err != nil {
 				logger.Error(err, "failed to delete alloy logs config")
@@ -444,7 +441,7 @@ func (r *ClusterMonitoringReconciler) reconcileDelete(ctx context.Context, clust
 			}
 		}
 
-		// Events-specific: Alloy events configuration - deployment that handles both kube event logs and traces
+		// alloy-event specific: Alloy events configuration - deployment that handles both kube event logs and traces - TODO rename alloy-events to alloy-cluster
 		if r.Config.Logging.IsLoggingEnabled(cluster) || r.Config.Tracing.IsTracingEnabled(cluster) {
 			// Clean up any existing alloy events configuration
 			err = r.AlloyEventsService.ReconcileDelete(ctx, cluster)